feat!: move `gix-odb::Write` trait to `gix-object::Write`.
There it's far more useful and plumbing crates are enabled to write
objects without pulling in `gix-odb` as dependency.
