+++ gix-odb/src/sink.rs
@@ -19,13 +19,13 @@ impl Sink {
     }
 }
 
-impl crate::traits::Write for Sink {
+impl gix_object::Write for Sink {
     fn write_stream(
         &self,
         kind: gix_object::Kind,
         mut size: u64,
         from: &mut dyn io::Read,
-    ) -> Result<gix_hash::ObjectId, crate::write::Error> {
+    ) -> Result<gix_hash::ObjectId, gix_object::write::Error> {
         let mut buf = [0u8; u16::MAX as usize];
         let header = gix_object::encode::loose_header(kind, size);
