+++ gix-odb/src/store_impls/dynamic/write.rs
@@ -24,11 +24,11 @@ pub use error::Error;
 
 use crate::store_impls::dynamic;
 
-impl<S> crate::Write for store::Handle<S>
+impl<S> gix_object::Write for store::Handle<S>
 where
     S: Deref<Target = dynamic::Store> + Clone,
 {
-    fn write_stream(&self, kind: Kind, size: u64, from: &mut dyn Read) -> Result<ObjectId, crate::write::Error> {
+    fn write_stream(&self, kind: Kind, size: u64, from: &mut dyn Read) -> Result<ObjectId, gix_object::write::Error> {
         let mut snapshot = self.snapshot.borrow_mut();
         Ok(match snapshot.loose_dbs.first() {
             Some(ldb) => ldb.write_stream(kind, size, from)?,
