+++ gix-odb/tests/odb/memory.rs
@@ -1,6 +1,6 @@
 use crate::odb::hex_to_id;
+use gix_object::Write;
 use gix_object::{tree, Exists, FindExt};
-use gix_odb::Write;
 use gix_testtools::tempfile::TempDir;
 
 #[test]
